% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcap_function.R
\name{carcap}
\alias{carcap}
\title{Carrying capacity function}
\usage{
carcap(a = 0.01, b = 0.01, c = 0.01, d = 0.01, H0, P0, t)
}
\arguments{
\item{t}{}
}
\description{
Carrying capacity function
}
\examples{

}
