% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_function.R
\name{sim}
\alias{sim}
\title{Simulation of predator and prey dynamics #1}
\usage{
sim(a = 0.01, b = 0.01, c = 0.01, d = 0.01, H0, P0, t)
}
\arguments{
\item{x}{this is the hare population size}

\item{y}{this is the lynx population size}

\item{z}{there is no z}
}
\value{
data.frame of hare and lynx sizes over time
}
\description{
Simulation of predator and prey dynamics #1
}
\examples{
sim()

}
